requirements.txt
python-telegram-bot==20.8
mysql-connector-python
python-dotenv
openai
streamlit
pandas
plotly